Use a default export for every component
Also remove index.js files to have unique file names, and remove the
default export from theme-legacy (not needed anymore).
