+++ src/components/BadgeBase/BadgePopoverBase.js
@@ -4,8 +4,8 @@ import { IconClose } from '../../icons'
 import { GU, textStyle } from '../../style'
 import { useTheme } from '../../theme'
 import { noop } from '../../utils'
-import { Button } from '../Button/Button'
-import { ButtonIcon } from '../Button/ButtonIcon'
+import Button from '../Button/Button'
+import ButtonIcon from '../Button/ButtonIcon'
 import Popover from '../Popover/Popover'
 import BadgePopoverActionType from './BadgePopoverActionType'
