+++ src/components/Box/Box.js
@@ -92,5 +92,4 @@ Box.propTypes = {
   ]),
 }
 
-export { Box }
 export default Box
