+++ src/components/Button/ButtonIcon.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { warnOnce } from '../../utils'
 import { GU } from '../../style'
-import { Button } from './Button'
+import Button from './Button'
 import ButtonBase from '../ButtonBase/ButtonBase'
 
 function ButtonIcon({ label, children, mode, ...props }) {
@@ -45,5 +45,4 @@ ButtonIcon.propTypes = {
   mode: PropTypes.oneOf(['button']),
 }
 
-export { ButtonIcon }
 export default ButtonIcon
