+++ src/components/DataView/DataView.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import { noop, warnOnce } from '../../utils'
 import { textStyle, GU } from '../../style'
 import { useTheme } from '../../theme'
-import { Box } from '../../components/Box/Box'
-import { Pagination } from '../../components/Pagination/Pagination'
+import Box from '../../components/Box/Box'
+import Pagination from '../../components/Pagination/Pagination'
 import { useLayout } from '../../components/Layout/Layout'
 import { usePublicUrl } from '../../providers/PublicUrl'
 import LoadingRing from '../LoadingRing/LoadingRing'
@@ -506,4 +506,4 @@ DataView.defaultProps = {
   status: 'default',
 }
 
-export { DataView }
+export default DataView
