+++ src/components/Info/Info.js
@@ -96,4 +96,12 @@ Info.propTypes = {
   borderColor: PropTypes.string,
 }
 
-export { Info }
+// Backward compatibility
+function Warning(props) {
+  return <Info mode="warning" {...props} />
+}
+Info.Action = Info
+Info.Permissions = Warning
+Info.Alert = Warning
+
+export default Info
