+++ src/components/Main/Main.js
@@ -6,7 +6,7 @@ import { Root } from '../../providers/Root'
 import { Viewport } from '../../providers/Viewport'
 import BaseStyles from '../BaseStyles/BaseStyles'
 import { ToastHub } from '../ToastHub/ToastHub'
-import { Layout } from '../Layout/Layout'
+import Layout from '../Layout/Layout'
 import { Theme } from '../../theme'
 import ScrollView from '../ScrollView/ScrollView'
 import { initContainsComponent } from '../../utils/contains-component'
