+++ src/components/Tabs/Tabs.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { noop, useInside, warnOnce } from '../../utils'
 import { useLayout } from '../Layout/Layout'
-import { Bar } from '../Bar/Bar'
+import Bar from '../Bar/Bar'
 import TabBarLegacy from './TabBarLegacy'
 import { TabsFullWidth } from './TabsFullWidth'
 import Tab from './Tab'
