+++ src/components/index.js
@@ -1,17 +1,10 @@
-export { Bar } from './Bar/Bar'
-export { DataView } from './DataView/DataView'
-export { DiscButton } from './DiscButton/DiscButton'
-export { Info } from './Info'
-export { Layout, useLayout } from './Layout/Layout'
-export { Link, ExternalLink, SafeLink } from './Link'
-export { Pagination } from './Pagination/Pagination'
-export { SearchInput } from './Input/SearchInput'
-export { Tag } from './Tag/Tag'
+export { ExternalLink, SafeLink } from './Link/LinkDeprecated'
 export { ToastHub, Toast, useToast } from './ToastHub/ToastHub'
 export { default as Accordion } from './Accordion/Accordion'
 export { default as AddressField } from './AddressField/AddressField'
 export { default as AppBadge } from './AppBadge/AppBadge'
 export { default as BackButton } from './BackButton/BackButton'
+export { default as Bar } from './Bar/Bar'
 export { default as BaseStyles } from './BaseStyles/BaseStyles'
 export { default as Box } from './Box/Box'
 export { default as Button } from './Button/Button'
@@ -24,7 +17,9 @@ export { default as Checkbox } from './Input/Checkbox'
 export { default as CircleGraph } from './CircleGraph/CircleGraph'
 export { default as ContextMenu } from './ContextMenu/ContextMenu'
 export { default as ContextMenuItem } from './ContextMenu/ContextMenuItem'
+export { default as DataView } from './DataView/DataView'
 export { default as DateRangePicker } from './DateRangePicker/DateRangePicker'
+export { default as DiscButton } from './DiscButton/DiscButton'
 export { default as Distribution } from './Distribution/Distribution'
 export { default as DropDown } from './DropDown/DropDown'
 export { default as EmptyStateCard } from './Card/EmptyStateCard'
@@ -36,11 +31,15 @@ export { default as FocusVisible } from './FocusVisible/FocusVisible'
 export { default as Header } from './Header/Header'
 export { default as Help } from './Help/Help'
 export { default as IdentityBadge } from './IdentityBadge/IdentityBadge'
+export { default as Info } from './Info/Info'
+export { default as Layout, useLayout } from './Layout/Layout'
 export { default as LineChart } from './LineChart/LineChart'
+export { default as Link } from './Link/Link'
 export { default as LoadingRing } from './LoadingRing/LoadingRing'
 export { default as Main, useContainsAppView } from './Main/Main'
 export { default as Modal } from './Modal/Modal'
 export { default as NavigationBar } from './NavigationBar/NavigationBar'
+export { default as Pagination } from './Pagination/Pagination'
 export { default as PartitionBar } from './PartitionBar/PartitionBar'
 export { default as Popover } from './Popover/Popover'
 export { default as ProgressBar } from './ProgressBar/ProgressBar'
@@ -50,6 +49,7 @@ export { default as RadioGroup } from './Radio/RadioGroup'
 export { default as RadioList } from './Radio/RadioList'
 export { default as RootPortal } from './RootPortal/RootPortal'
 export { default as ScrollView } from './ScrollView/ScrollView'
+export { default as SearchInput } from './Input/SearchInput'
 export {
   default as SidePanel,
   useSidePanel,
@@ -66,6 +66,7 @@ export { default as TableCell } from './Table/TableCell'
 export { default as TableHeader } from './Table/TableHeader'
 export { default as TableRow } from './Table/TableRow'
 export { default as Tabs } from './Tabs/Tabs'
+export { default as Tag } from './Tag/Tag'
 export { default as TextCopy } from './TextCopy/TextCopy'
 export { default as TextInput } from './Input/TextInput'
 export { default as Timer } from './Timer/Timer'
