+++ src/theme-legacy/index.js
@@ -1,1 +1,1 @@
-export { theme as default, theme, themeDark, brand, colors } from './palettes'
+export { theme, themeDark, brand, colors } from './palettes'
